pyproject.toml
[project]
name = "finanzonline_uid"
version = "2.0.0"
description = "Python library and CLI for querying Level 2 UID checks via FinanzOnline Austria"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rich-click>=1.9.4",
    "lib_cli_exit_tools>=2.2.3",
    "lib_log_rich>=6.0.0",
    "lib_layered_config>=5.1.0",
    "btx_lib_mail>=1.0.3",
    "zeep>=4.3.2",
    "filelock>=3.20.1",
    "orjson>=3.11.5",
]
license = { text = "MIT" }
authors = [{ name = "bitranox", email = "bitranox@gmail.com" }]
keywords = ["uid", "vat", "austria", "finanzonline", "tax", "verification", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Environment :: Console",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/bitranox/finanzonline_uid"
Repository = "https://github.com/bitranox/finanzonline_uid.git"
Issues = "https://github.com/bitranox/finanzonline_uid/issues"

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "pytest-cov>=7.0.0",
  "ruff>=0.14.10",
  "pyright>=1.1.407",
  "bandit>=1.9.2",
  "build>=1.3.0",
  "twine>=6.2.0",
  "codecov-cli>=11.2.6",
  "pip-audit>=2.10.0",
  "textual>=6.11.0",
  "import-linter>=2.9",
  "rtoml>=0.13.0",
]

[project.scripts]
finanzonline_uid = "finanzonline_uid.cli:main"
finanzonline-uid = "finanzonline_uid.cli:main"

[build-system]
requires = ["hatchling>=1.28.0"]
build-backend = "hatchling.build"



[tool.hatch.build.targets.wheel]
packages = ["src/finanzonline_uid"]
include = [
  "src/finanzonline_uid/py.typed",
  "src/finanzonline_uid/defaultconfig.toml",
  "src/finanzonline_uid/defaultconfig.d/*.toml",
  "src/finanzonline_uid/locales/**/*.mo",
]

[tool.ruff]
line-length = 160
target-version = "py310"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
# ignore the "unused import sys" in the jupyter notebook - which we need (it is a linter hickup)
"notebooks/*.ipynb" = ["F401"]

[tool.pyright]
typeCheckingMode = "strict"
extraPaths = [
  "src",
]
exclude = [
  "scripts/menu.py",
  # zeep SOAP library has no type stubs, causing many reportUnknownMemberType errors
  "src/finanzonline_uid/adapters/finanzonline/session_client.py",
  "src/finanzonline_uid/adapters/finanzonline/uid_query_client.py",
]

[tool.pytest.ini_options]
addopts = ["--doctest-modules"]
doctest_optionflags = ["ELLIPSIS", "NORMALIZE_WHITESPACE"]
pythonpath = ["src"]
markers = [
  "os_agnostic: test runs on every supported operating system",
  "os_windows: test exercises Windows-only behavior",
  "os_macos: test exercises macOS-only behavior",
  "os_posix: test exercises behavior requiring POSIX semantics",
  "os_linux: test exercises Linux-only behavior",
]

[tool.coverage.run]
branch = true
parallel = false
data_file = ".coverage.unit"
source = ["src/finanzonline_uid"]

[tool.coverage.report]
omit = [
  "tests/*",
]
fail_under = 85
show_missing = true

[tool.coverage.json]
output = "coverage.json"
pretty_print = true

[tool.clean]
# Patterns to remove when running `make clean` or `python -m scripts clean`
patterns = [
  ".hypothesis",
  ".import_linter_cache",
  ".pytest_cache",
  ".ruff_cache",
  ".pyright",
  ".mypy_cache",
  ".tox",
  ".nox",
  ".eggs",
  "*.egg-info",
  "build",
  "dist",
  "htmlcov",
  ".coverage",
  "coverage.xml",
  "codecov.sh",
  ".cache",
  "result",
]

[tool.git]
# Default git remote for push/release commands
default-remote = "origin"


[tool.scripts.test]
# Configuration for scripts/test.py test runner
pytest-verbosity = "-vv"
coverage-report-file = "coverage.xml"
src-path = "src"
[tool.importlinter]
root_package = "finanzonline_uid"

[[tool.importlinter.contracts]]
name = "CLI depends on behaviors only"
type = "layers"
layers = [
  "finanzonline_uid.cli",
  "finanzonline_uid.behaviors",
]

[tool.pip-audit]
# Known vulnerabilities to ignore (unfixable or accepted risk)
# GHSA-4xh5-x5gv-qwph: py library ReDoS (no fix available, low risk)
# PYSEC-2022-42969: same as above (alias)
# PYSEC-2022-43012, PYSEC-2025-49, CVE-2024-6345: setuptools vulns (CI uses old version)
ignore-vulns = ["GHSA-4xh5-x5gv-qwph", "PYSEC-2022-42969", "PYSEC-2022-43012", "PYSEC-2025-49", "CVE-2024-6345", "CVE-2025-68146"]
